# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Api",
        "description": "Demo Car Rental Api",
        "version": "1.0"
    },
    "paths": {
...
        "/car-rental/api/v1/cars/{id}": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Get car details",
                "operationId": "d521feef0acba1c9e89afb4ea705310c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "AuthorizationHeader",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer <token>",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "vehicle-id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "vehicle-name": {
                                            "type": "string",
                                            "example": "Toyota Hybrid"
                                        },
                                        "transmission": {
                                            "type": "string",
                                            "example": "automatic"
                                        },
                                        "daily-rate": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "3.8"
                                        },
                                        "seats": {
                                            "type": "number",
                                            "example": "5"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://base-url/path/to/image"
                                        },
                                        "available": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car-rental/api/v1/cars": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Get cars list",
                "operationId": "b2e6d36ec2df569ba7729591cd7858c7",
                "parameters": [
                    {
                        "parameter": "AuthorizationHeader",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer <token>",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "vehicle-id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "vehicle-name": {
                                                        "type": "string",
                                                        "example": "Toyota Hybrid"
                                                    },
                                                    "transmission": {
                                                        "type": "string",
                                                        "example": "automatic"
                                                    },
                                                    "daily-rate": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "3.8"
                                                    },
                                                    "seats": {
                                                        "type": "number",
                                                        "example": "5"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://base-url/path/to/image"
                                                    },
                                                    "available": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}